refactor: cart and product actions
